Mejoras visuales en musicnav
