+++ src/Static/CSS/InstrumentSelector.css
@@ -1,0 +1,13 @@
+#Instrument-Container{
+	margin: 0 !important;
+	padding: 0 !important;
+}
+#Prev-Container{
+	padding-right: 0 !important;
+}
+#Next-Container{
+	padding-left: 0 !important;
+}
+#Instrument{
+	padding: 0 !important;
+}
